[UPDATE] change structure to get secrets
