+++ modules/museums/find_museum/connect_db.py
@@ -1,0 +1,39 @@
+import json
+import boto3
+import psycopg2
+
+def  get_db_connection():
+    secrets = get_secrets()
+    host = secrets['POSTGRES_HOST']
+    user = 'default'
+    password = secrets['POSTGRES_PASSWORD']
+    database = secrets['POSTGRES_DATABASE']
+    return psycopg2.connect(
+       host= host,
+       user= user,
+       password= password,
+       database= database
+   )
+
+
+def get_secrets():
+ secret_name = "prod/musepa/vercel/postgres"
+    region_name = "us-west-1"
+
+    # Create a Secrets Manager client
+    session = boto3.session.Session()
+    client = session.client(
+        service_name='secretsmanager',
+        region_name=region_name
+    )
+
+    try:
+        get_secret_value_response = client.get_secret_value(
+            SecretId=secret_name
+        )
+    except Exception as e:
+        raise e
+
+    secret = get_secret_value_response['SecretString']
+    return json.loads(secret)
+
