+++ modules/museums/find_museum/functions.py
@@ -1,6 +1,3 @@
-import json
-import boto3
-import psycopg2
 from datetime import datetime, date
 
 
@@ -9,40 +6,3 @@ def datetime_serializer(obj):
     if isinstance(obj, (datetime, date)):
         return obj.isoformat()
     raise TypeError(f"Type {type(obj)} not serializable")
-
-
-def  get_db_connection():
-    secrets = get_secrets()
-    host = secrets['POSTGRES_HOST']
-    user = 'default'
-    password = secrets['POSTGRES_PASSWORD']
-    database = secrets['POSTGRES_DATABASE']
-    return psycopg2.connect(
-       host= host,
-       user= user,
-       password= password,
-       database= database
-   )
-
-
-def get_secrets():
- secret_name = "prod/musepa/vercel/postgres"
-    region_name = "us-west-1"
-
-    # Create a Secrets Manager client
-    session = boto3.session.Session()
-    client = session.client(
-        service_name='secretsmanager',
-        region_name=region_name
-    )
-
-    try:
-        get_secret_value_response = client.get_secret_value(
-            SecretId=secret_name
-        )
-    except Exception as e:
-        raise e
-
-    secret = get_secret_value_response['SecretString']
-    return json.loads(secret)
-
